ai_support_agent/requirement.txt
gradio==4.19.2
openai==1.12.0
asyncio
pydantic==2.6.1
python-dotenv==1.0.1
aiofiles==23.2.1